Global update
Global restructurization of subroutines and functions.
Provided three-body case for processing.
Provided keys in global parameters for setting the program.
Libration classifier is more smart and can ignore the most part of slow circulations (but still needs calibration, I think).
This version is stable and almost ready for a work.
Unit tests will be provided with next commit.