Technical update
Found that logical core existance in processor should be taken into account.
Improved the classifier: now cross-periodigram is not checked on periods close to orbital periods of objects.
Fixed the bug in function "count_axis_3body" and prepared a pattern for the next update (still need to get second expression for precession rate).